requirements.txt
fastapi[standard]
uvicorn
PyMySQL
SQLAlchemy
cryptography
python-multipart
pyjwt
pwdlib[argon2]
Pillow
imagehash
